+++ lib/src/segment_style/hex_segment_style.dart
@@ -14,16 +14,11 @@ class HexSegmentStyle extends SegmentStyle {
 
   /// Creates elongated hexagon [SegmentStyle].
   const HexSegmentStyle({
-    Size? segmentBaseSize,
-    Color? enabledColor,
-    Color? disabledColor,
+    super.segmentBaseSize,
+    super.enabledColor,
+    super.disabledColor,
     this.segmentSpacing = 4.0,
-  })  : _halfSpace = segmentSpacing / 2,
-        super(
-          segmentBaseSize: segmentBaseSize,
-          enabledColor: enabledColor,
-          disabledColor: disabledColor,
-        );
+  }) : _halfSpace = segmentSpacing / 2;
 
   @override
   Path createHorizontalPath(SegmentPosition position, Size segmentSize) {